Zwischenstand

+++ ibclient/covcall.py
@@ -4,13 +4,13 @@ from datetime import datetime
 
 class covered_call():
 
-    def __init__(self, right: str, underlyer: Contract, strike: int, expiry: datetime):
+    def __init__(self, right: str, underlyer: Contract, strike: float, expiry: datetime):
         self.ul = underlyer
         self.strike = strike
         self.right = right
         self.expiry = expiry
-        self.stkprice = 0
-        self.optprice = 0
+        self.stkprice = 0.0
+        self.optprice = 0.0
 
     def set_stk_price(self,price):
         self.stkprice = price
@@ -21,8 +21,8 @@ class covered_call():
     def getTimevalue(self):
         if self.stkprice > self.strike:
             #ITM
-            intrinsic_val = self.stkprice - self.strike
-            return self.optprice - intrinsic_val
+            intrinsic_val = float(self.stkprice) - float(self.strike)
+            return float(self.optprice) - intrinsic_val
         else:
             return self.optprice
